Fixed websocket connections maybe
