+++ backend/KafkaAuction/Models/Auction.cs
@@ -9,7 +9,7 @@ public record Auction
     public required string Title { get; set; }
     public string? Description { get; set; }
     public required decimal Starting_Price { get; set; }
-    public decimal? Current_Price { get; set; } = 0;
+    public required decimal Current_Price { get; set; } = 0;
     public int? Number_Of_Bids { get; set; } = 0;
     public string? Winner { get; set; }
     public string Created_At { get; set; } = DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss");
