+++ backend/KafkaAuction/Services/Interfaces/WebSocketService.cs/IWebSocketHandler.cs
@@ -6,4 +6,5 @@ namespace KafkaAuction.Services.Interfaces.WebSocketService;
 public interface IWebSocketHandler
 {
     Task HandleWebSocketAsync(HttpContext context, WebSocket webSocket, string auctionId, WebPages page);
+    int GetActiveConnectionsCount();
 }
