+++ backend/KafkaAuction/Services/WebSocketService/WebSocketHandler.cs
@@ -1,3 +1,4 @@
+using System.Collections.Concurrent;
 using System.Net.WebSockets;
 using KafkaAuction.Enums;
 using KafkaAuction.Services.Interfaces.WebSocketService;
@@ -9,33 +10,76 @@ public class WebSocketHandler : IWebSocketHandler
     private readonly ILogger<WebSocketHandler> _logger;
     private readonly IAuctionWebSocketService _auctionWebSocketService;
     private readonly IChatWebSocketService _chatWebSocketService;
+    private readonly ConcurrentDictionary<WebSocket, bool> _connections;
 
     public WebSocketHandler(ILogger<WebSocketHandler> logger, IAuctionWebSocketService auctionWebSocketService, IChatWebSocketService chatWebSocketService)
     {
         _logger = logger;
         _auctionWebSocketService = auctionWebSocketService;
         _chatWebSocketService = chatWebSocketService;
+        _connections = new ConcurrentDictionary<WebSocket, bool>();
     }
 
     public async Task HandleWebSocketAsync(HttpContext context, WebSocket webSocket, string auctionId, WebPages page)
     {
+        _connections.TryAdd(webSocket, true);
         _logger.LogInformation("WebSocket connection established for {Page} with auctionId {AuctionId}.", page, auctionId);
 
-        if (page == WebPages.AuctionOverview)
+        try
         {
-            await _auctionWebSocketService.SubscribeToAuctionOverviewUpdatesAsync(webSocket);
+            if (page == WebPages.AuctionOverview)
+            {
+                await _auctionWebSocketService.SubscribeToAuctionOverviewUpdatesAsync(webSocket);
+            }
+            else if (page == WebPages.SpesificAuction)
+            {
+                await _auctionWebSocketService.SubscribeToAuctionBidUpdatesAsync(webSocket, auctionId);
+            }
+            else if (page == WebPages.Chat)
+            {
+                await _chatWebSocketService.SubscribeToChatMessagesForAuctionAsync(webSocket, auctionId);
+            }
+            else
+            {
+                throw new InvalidOperationException("Invalid page");
+            }
+
+            // Keep the WebSocket open until closed by the client
+            var buffer = new byte[1024 * 4];
+            WebSocketReceiveResult? result;
+            do
+            {
+                result = await ReceiveWebSocketMessageAsync(webSocket, buffer);
+            } while (result != null && !result.CloseStatus.HasValue);
         }
-        else if (page == WebPages.SpesificAuction)
+        finally
         {
-            await _auctionWebSocketService.SubscribeToAuctionBidUpdatesAsync(webSocket, auctionId);
+            _connections.TryRemove(webSocket, out _);
+            if (webSocket.State == WebSocketState.Open)
+            {
+                await webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "Closed by server", CancellationToken.None);
+            }
         }
-        else if (page == WebPages.Chat)
+    }
+
+    private async Task<WebSocketReceiveResult?> ReceiveWebSocketMessageAsync(WebSocket webSocket, byte[] buffer)
+    {
+        if (webSocket.State != WebSocketState.Open)
+            return null;
+
+        try
         {
-            await _chatWebSocketService.SubscribeToChatMessagesForAuctionAsync(webSocket, auctionId);
+            return await webSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
         }
-        else
+        catch (WebSocketException ex)
         {
-            throw new InvalidOperationException("Invalid page");
+            _logger.LogError(ex, "Error receiving WebSocket message");
+            return null;
         }
+    }
+
+    public int GetActiveConnectionsCount()
+    {
+        return _connections.Count;
     }
 }
