+++ frontend/src/components/debugComponents/CheckStreams.vue
