+++ frontend/src/components/debugComponents/DropStreams.vue
