+++ frontend/src/components/debugComponents/InsertAuction.vue
